# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        example: session not found
        type: string
    type: object
...
  dto.SendResponse:
    properties:
      messageId:
        example: ABCD1234
        type: string
      timestamp:
        example: 1699999999
        type: integer
    type: object
  dto.SendStatusRequest:
    properties:
      caption:
        example: Status caption
        type: string
      image:
        example: base64_encoded_image
        type: string
      mimetype:
        example: image/jpeg
        type: string
      text:
        example: My status update
        type: string
      video:
        example: base64_encoded_video
        type: string
    type: object
...
  dto.StatusPrivacyResponse:
    properties:
      privacy: {}
    type: object
...
info:
  contact: {}
  description: WhatsApp API with Chatwoot Integration
  title: ZPWoot WhatsApp API
  version: 0.1.0
paths:
...
  /sessions/{sessionId}/stories:
    post:
      consumes:
      - application/json
      description: Post a text or media story (status update visible to contacts)
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Story data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Post a story
      tags:
      - status
  /sessions/{sessionId}/stories/privacy:
    get:
      description: Get who can see your status updates
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StatusPrivacyResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get status privacy settings
      tags:
      - status
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"
...